Return non-zero if any errors occur
Builds currently succeed whenever using the `--download-artifacts` flag,
because `exit_code` is overridden with the return value of `curl`.

This initiates `exit_code=0` and ensures that any errors will be
accumulated to it as the build progresses. If any errors occur, either
during the build or the downloading of artifacts, `exit_code` will have
a non-zero value, which will ensure the script exits with an error code.